Rename Event -> PolymerEvent
