+++ template/ElementEvent.template
@@ -8,7 +8,7 @@ import jsinterop.annotations.JsProperty;
 import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
 import elemental2.dom.Element;
-import com.vaadin.polymer.elemental.Event;
+import com.vaadin.polymer.PolymerEvent;
 import elemental2.dom.EventInit;
 import elemental2.dom.EventTarget;
 
@@ -16,7 +16,7 @@ import elemental2.dom.EventTarget;
  * <%= getDescription(' ') %>
  */
 @JsType(isNative=true, namespace=JsPackage.GLOBAL, name="Event")
-public class <%= camelCase(name) %>Event extends Event {
+public class <%= camelCase(name) %>Event extends PolymerEvent {
 
     @JsOverlay public static final String NAME = "<%= computeName(name) %>";
 
@@ -34,7 +34,7 @@ public class <%= camelCase(name) %>Event extends Event {
     public native Detail getDetail();
 
     @JsType(isNative=true)
-    public interface Detail extends Event.Detail {
+    public interface Detail extends PolymerEvent.Detail {
 <% _.forEach(params, function(param){ %>
         /**
          * <%= getDescription('         ', param) %>
